+++ src/renderer/domains/app/deep-link/index.ts
@@ -1,0 +1,2 @@
+export { deepLinkService, createDeepLinkHandler, extractParams, processDeepLinkChain } from './service';
+export type { DeepLinkHandler, HandlerResult, DeepLinkHandlerConfig } from './service';
