+++ src/renderer/domains/app/index.ts
@@ -1,0 +1,1 @@
+export * from './deep-link';
