--- src/renderer/shared/lib/deep-link/index.ts
@@ -1,2 +1,0 @@
-export { deepLinkService, createDeepLinkHandler, extractParams, processDeepLinkChain } from './deep-link-service';
-export type { DeepLinkHandler, HandlerResult, DeepLinkHandlerConfig } from './deep-link-service';
